% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reefFishDiversity}
\alias{reefFishDiversity}
\title{Reef fish diversity}
\format{
A data frame with 55 rows and 6 columns. Each row corresponds to a different family and the variables are:
\describe{
  \item{Family}{Reference number to original article (see source)}
  \item{Phyl_name}{String containing family and species names}
  \item{Spp_rich}{Species richness (number of species within the family)}
  \item{Divergence}{???}
  \item{Egg}{Reproductive strategy with respect to eggs}
  \item{Egg_3}{???}
}
}
\source{
Riginos, C., Buckley, Y.M., Blomberg, S.P. & Treml, E.A. 2014.
The American Naturalist 184: 52-64 (2014)
}
\usage{
reefFishDiversity
}
\description{
A dataset containing data on species richness and divergence of 55 families of reef fishes.
}
\keyword{datasets}
